Split tests within expectation
